requirements.txt
Pillow==5.3.0
mysqlclient==1.3.14